chore: remove suite
